second commit

+++ src/TP1/Statisticable.java
@@ -1,5 +1,5 @@
 package TP1;
 
 public interface Statisticable {
-	public abstract double getValue();
+	public abstract float getValue();
 }
